requirements.txt
fastapi==0.61.0
hashids==1.3.1
paho-mqtt==1.5.0
tinydb==4.1.1
pytest==6.0.1
requests==2.24.0
